Preparation work to support LowCardinality columns
