+++ ClickHaskell/ClickHaskell.hs
@@ -15,13 +15,16 @@ module ClickHaskell
 
   -- * Reading and writing
   , Table
-  , View, parameter, Parameter
+  , Column
 
   -- ** Reading
   , ReadableFrom(..)
   , select
   , selectFrom
+
+  -- *** Reading from view
   , selectFromView
+  , View, parameter, Parameter
 
   -- ** Writing
   , WritableInto(..)
@@ -41,15 +44,14 @@ import ClickHaskell.NativeProtocol
   , ServerPacketType(..), HelloResponse(..), ExceptionPacket, latestSupportedRevision
   )
 import ClickHaskell.Versioning (ProtocolRevision)
-import ClickHaskell.Columns (HasColumns (..), KnownColumns (..), appendColumn, emptyColumns, mkColumn)
+import ClickHaskell.Columns (HasColumns (..), KnownColumns (..), rowsCount, WritableInto (..), ReadableFrom (..))
 import ClickHaskell.Parameters (Parameter, parameter, parameters, Parameters, CheckParameters)
 import ClickHaskell.DeSerialization (Serializable(..), Deserializable(..), DeserializableColumns(..))
 
 -- GHC included
 import Control.Exception (Exception, SomeException, bracketOnError, catch, finally, throwIO)
-import Data.Bifunctor (bimap)
 import Data.Binary.Get (Decoder (..), Get, runGetIncremental)
-import Data.ByteString.Builder (Builder, byteString, toLazyByteString)
+import Data.ByteString.Builder (byteString, toLazyByteString)
 import Data.ByteString.Char8 as BS8 (fromStrict, pack)
 import Data.ByteString.Lazy.Internal as BL (ByteString (..), LazyByteString)
 import Data.Int (Int64)
@@ -57,7 +59,6 @@ import Data.Kind (Type)
 import Data.Maybe (listToMaybe)
 import Data.Text (Text)
 import Data.Typeable (Proxy (..))
-import GHC.Generics
 import GHC.Stack (HasCallStack, callStack, prettyCallStack)
 import GHC.TypeLits (KnownSymbol, Symbol, symbolVal)
 import System.Timeout (timeout)
@@ -148,6 +149,11 @@ ping conn@MkConnection{sock, revision} = do
 
 -- * Querying
 
+data Table (name :: Symbol) (columns :: [Type])
+
+instance HasColumns (Table name columns) where
+  type GetColumns (Table _ columns) = columns
+
 -- ** Selecting
 
 selectFrom ::
@@ -183,6 +189,13 @@ select conn@MkConnection{sock, user, revision} query = do
     )
   handleSelect @(Columns columns) conn emptyBuffer
 
+
+-- *** View
+
+instance HasColumns (View name columns parameters) where
+  type GetColumns (View _ columns _) = columns
+
+data View (name :: Symbol) (columns :: [Type]) (parameters :: [Type])
 
 selectFromView ::
   forall view record name columns parameters passedParameters
@@ -222,6 +235,7 @@ handleSelect conn previousBuffer = do
     EndOfStream         -> pure []
     Exception exception -> throwIO (DatabaseException exception)
     otherPacket         -> throwIO (ProtocolImplementationError $ UnexpectedPacketType otherPacket)
+
 
 -- ** Inserting
 
@@ -265,6 +279,9 @@ handleInsertResult conn@MkConnection{..} buffer records = do
     EndOfStream         -> pure ()
     Exception exception -> throwIO (DatabaseException exception)
     otherPacket         -> throwIO (ProtocolImplementationError $ UnexpectedPacketType otherPacket)
+
+
+
 
 -- * Reading
 
@@ -323,180 +340,3 @@ data ConnectionError
   = NoAdressResolved
   | EstablishTimeout
   deriving (Show, Exception)
-
-
-
-
--- * Interface
-
-data View (name :: Symbol) (columns :: [Type]) (parameters :: [Type])
-
-data Table (name :: Symbol) (columns :: [Type])
-
-
-instance HasColumns (Table name columns)
-  where
-  type GetColumns (Table _ columns) = columns
-
-instance HasColumns (View name columns parameters)
-  where
-  type GetColumns (View _ columns _) = columns
-
--- ** Reading
-
-type GenericReadable record hasColumns =
-  ( Generic record
-  , GReadable (GetColumns hasColumns) (Rep record)
-  )
-
-class
-  ( HasColumns hasColumns
-  , DeserializableColumns (Columns (GetColumns hasColumns))
-  ) =>
-  ReadableFrom hasColumns record
-  where
-  default fromColumns :: GenericReadable record hasColumns => Columns (GetColumns hasColumns) -> [record]
-  fromColumns :: Columns (GetColumns hasColumns) -> [record]
-  fromColumns = map to . gFromColumns @(GetColumns hasColumns)
-
-  default readingColumns :: GenericReadable record hasColumns => Builder
-  readingColumns :: Builder
-  readingColumns = gReadingColumns @(GetColumns hasColumns) @(Rep record)
-
-
-class GReadable columns f
-  where
-  gFromColumns :: Columns columns -> [f p]
-  gReadingColumns :: Builder
-
-instance
-  GReadable columns f
-  =>
-  GReadable columns (D1 c (C1 c2 f))
-  where
-  {-# INLINE gFromColumns #-}
-  gFromColumns = map (M1 . M1) . gFromColumns @columns
-  gReadingColumns = gReadingColumns @columns @f
-
-instance
-  GReadable columns (left :*: (right1 :*: right2))
-  =>
-  GReadable columns ((left :*: right1) :*: right2)
-  where
-  {-# INLINE gFromColumns #-}
-  gFromColumns rev = do
-    (l :*: (r1 :*: r2)) <- gFromColumns rev
-    pure ((l :*: r1) :*: r2)
-  gReadingColumns = gReadingColumns @columns @(left :*: (right1 :*: right2))
-
-instance
-  ( KnownColumn (Column name chType)
-  , FromChType chType inputType
-  , GReadable restColumns right
-  )
-  =>
-  GReadable
-    (Column name chType ': restColumns)
-    (S1 (MetaSel (Just name) a b f) (Rec0 inputType) :*: right)
-  where
-  {-# INLINE gFromColumns #-}
-  gFromColumns (AddColumn column extraColumns) =
-    zipWith (:*:)
-      (gFromColumns (AddColumn column emptyColumns))
-      (gFromColumns extraColumns)
-  gReadingColumns =
-    renderColumnName @(Column name chType)
-    <> ", " <> gReadingColumns @restColumns @right
-
-instance
-  ( KnownColumn (Column name chType)
-  , FromChType chType inputType
-  ) => GReadable '[Column name chType] ((S1 (MetaSel (Just name) a b f)) (Rec0 inputType))
-  where
-  {-# INLINE gFromColumns #-}
-  gFromColumns (AddColumn (MkColumn _ column) _) = map (M1 . K1 . fromChType @chType) column
-  gReadingColumns = renderColumnName @(Column name chType)
-
-
--- ** Writing
-
-type GenericWritable record hasColumns =
-  ( Generic record
-  , GWritable (GetColumns hasColumns) (Rep record)
-  )
-
-class
-  ( HasColumns hasColumns
-  , KnownColumns (Columns (GetColumns hasColumns))
-  , Serializable (Columns (GetColumns hasColumns))
-  , DeserializableColumns (Columns (GetColumns hasColumns))
-  )
-  =>
-  WritableInto hasColumns record
-  where
-  default toColumns :: GenericWritable record hasColumns => UVarInt -> [record] -> Columns (GetColumns hasColumns)
-  toColumns :: UVarInt -> [record] -> Columns (GetColumns hasColumns)
-  toColumns size = gToColumns @(GetColumns hasColumns) size . map from
-
-  default writingColumns :: GenericWritable record hasColumns =>  Builder
-  writingColumns :: Builder
-  writingColumns = gWritingColumns @(GetColumns hasColumns) @(Rep record)
-
-
-class GWritable columns f
-  where
-  gToColumns :: UVarInt -> [f p] -> Columns columns
-  gWritingColumns :: Builder
-
-instance
-  GWritable columns f
-  =>
-  GWritable columns (D1 c (C1 c2 f))
-  where
-  {-# INLINE gToColumns #-}
-  gToColumns size = gToColumns @columns size . map (unM1 . unM1)
-  {-# INLINE gWritingColumns #-}
-  gWritingColumns = gWritingColumns @columns @f
-
-instance
-  GWritable columns (left1 :*: (left2 :*: right))
-  =>
-  GWritable columns ((left1 :*: left2) :*: right)
-  where
-  {-# INLINE gToColumns #-}
-  gToColumns size = gToColumns size . map (\((l1 :*: l2) :*: r) -> l1 :*: (l2 :*: r))
-  {-# INLINE gWritingColumns #-}
-  gWritingColumns = gWritingColumns @columns @(left1 :*: (left2 :*: right))
-
-instance
-  ( GWritable '[Column name chType] (S1 (MetaSel (Just name) a b f) (Rec0 inputType))
-  , GWritable restColumns right
-  , KnownColumn (Column name chType)
-  , ToChType chType inputType
-  )
-  =>
-  GWritable (Column name chType ': restColumns) (S1 (MetaSel (Just name) a b f) (Rec0 inputType) :*: right)
-  where
-  {-# INLINE gToColumns #-}
-  gToColumns size
-    = uncurry appendColumn
-    . bimap (mkColumn size) (gToColumns @restColumns size)
-    . unzip
-    . map (\(l :*: r) -> ((toChType . unK1 . unM1) l, r))
-     
-  {-# INLINE gWritingColumns #-}
-  gWritingColumns =
-    renderColumnName @(Column name chType)
-    <> ", " <> gWritingColumns @restColumns @right
-
-instance
-  ( ToChType chType inputType
-  , KnownColumn (Column name chType)
-  )
-  =>
-  GWritable '[Column name chType] (S1 (MetaSel (Just name) a b f) (Rec0 inputType))
-  where
-  {-# INLINE gToColumns #-}
-  gToColumns size rows = (MkColumn size . map (toChType . unK1 . unM1)) rows `appendColumn` emptyColumns
-  {-# INLINE gWritingColumns #-}
-  gWritingColumns = renderColumnName @(Column name chType)
