+++ ClickHaskell/ClickHaskell/DeSerialization.hs
@@ -23,7 +23,7 @@ import Data.ByteString.Builder (Builder, word8)
 import Data.Coerce (coerce)
 import Data.Typeable (Proxy (..))
 import GHC.Generics
-import GHC.TypeLits (ErrorMessage (..), KnownNat, KnownSymbol, TypeError, natVal)
+import GHC.TypeLits (ErrorMessage (..), KnownNat, TypeError, natVal)
 
 -- * Deserialization
 
@@ -285,31 +285,31 @@ instance
 instance
   ( KnownColumn (Column name chType)
   , IsChType chType
-  , KnownSymbol name
   , Serializable chType
   ) => Serializable (Column name chType) where
   {-# INLINE serialize #-}
-  serialize rev (MkColumn _size values)
+  serialize rev column
     =  serialize rev (toChType @ChString $ renderColumnName @(Column name chType))
     <> serialize rev (toChType @ChString $ renderColumnType @(Column name chType))
     -- serialization is not custom
     <> afterRevision @DBMS_MIN_REVISION_WITH_CUSTOM_SERIALIZATION rev (serialize @ChUInt8 rev 0)
-    <> mconcat (Prelude.map (serialize @chType rev) values)
+    <> mconcat (Prelude.map (serialize @chType rev) (columnValues column))
 
 instance {-# OVERLAPPING #-}
   ( KnownColumn (Column name (Nullable chType))
   , IsChType chType
-  , KnownSymbol name
   , Serializable chType
   ) => Serializable (Column name (Nullable chType)) where
   {-# INLINE serialize #-}
-  serialize rev (MkColumn _size values)
+  serialize rev column
     =  serialize rev (toChType @ChString $ renderColumnName @(Column name (Nullable chType)))
     <> serialize rev (toChType @ChString $ renderColumnType @(Column name (Nullable chType)))
     -- serialization is not custom
     <> afterRevision @DBMS_MIN_REVISION_WITH_CUSTOM_SERIALIZATION rev (serialize @ChUInt8 rev 0)
-    {- Nulls -} <> mconcat (Prelude.map (serialize @ChUInt8 rev . \case Nothing -> 1; Just _ -> 0) values)
-    {- Values -} <> mconcat (Prelude.map (maybe (serialize @chType rev defaultValueOfTypeName) (serialize @chType rev)) values)
+    -- Nulls
+    <> mconcat (Prelude.map (serialize @ChUInt8 rev . maybe 1 (const 0)) (columnValues column))
+    -- Values
+    <> mconcat (Prelude.map (serialize @chType rev . maybe defaultValueOfTypeName id) (columnValues column))
 
 
 -- ** Generics
